t240901a
+ Add skills
+ Add fireball skill
+ Add weapon skill
+ Add Scenes in order to store events of buttons and store something
* Fixed bug: unit's position after ending battle
+ Add attributes
! Attack, HP, Mana, Stamina will base on attributes
+ Add human's cast animation

+++ Assets/_Scripts/Managers/BattleGameManager.cs
@@ -534,11 +534,11 @@ public class BattleGameManager : Singleton<BattleGameManager>
                     }
                     else if (matchedItem.tag == "Gold")
                     {
-                        _currentPlayer.IncreaseGold((int)_bonusFactor);
+                        _currentPlayer.IncreaseGold(_bonusFactor);
                     }
                     else if (matchedItem.tag == "Exp")
                     {
-                        _currentPlayer.IncreaseExp((int)_bonusFactor);
+                        _currentPlayer.IncreaseExp(_bonusFactor);
                     }
                 }
                 DestroyItem(matchedItem);
